Change instruction handling for extended instructions
These are now handled by the opcode mechanism as well instead of being lumped together under the catchall CB opcode. The opcode struct now contains the opcode value and the information if the opcode is a CB opcode (extended opcode).
Since fetching an instruction now sets the program counter to point past the fetched instructions, incrementeng the program counter is only required for instructions which read immediate data. It was removed for all other instructions