Remove unused code
This is from the previous implementation where instruction handling was
done in two different ways. Either directly in the cpu class or with
instructions as separate classes. This was refactored to have every
instruction as a member function of the CPU class for consistency.
The other approach was far move verbose while not providing many
benefits.