Bugfixes in ADD instructions
The word add instructions dont modify the zero flag but take one cycle
longer than the byte instructions.
The half carry for the word instructions is calculated from the high
byte, not the low byte.
While the incoming value of data will only be 8 bits, inside the
function data is used to store up to 16 bits, so the parameter has to be
changed.


Former-commit-id: 65e59ddc84a9c5232dad90994fcc86d5c8348daa