Add blargg4 test rom
Refactor copying the test data files to the build folder by making it
only one custom target on which the unit tests depend. This means adding
test data wont require modifying the CMakeLists.txt anymore and test
data will be copied even if the unittests wont need to be build.


Former-commit-id: 7731975635432d07f06d8c26cc8f8f844b90bd82 [formerly 85182a63d1ecec15cc77b686d9a9c9155968ee3f]
Former-commit-id: 211f770b48e72e6b72e911a879642a3472fdc4e0