Implement battery backed cartridge RAM
Use a file to store cartridge RAM data and load it again when emulator
is restarted. This means saving within the game is supported.
Use a memory mapped file to directly write to this file instead of
storing a second buffer. Flush the file every frame to avoid losing data
on crash.