Improve PPU sprite drawing
Previously all sprites were drawn in the VBlank state, not line by line
such as on the original Game Boy.
The 10-sprite-per-line-limit was not taken into account.
Also draw all debug windows at the begin of Vblank instead when leaving
it so changes made during vblank are not reflected in the debug windows.