Trigger interrupts after writing the line
Otherwise the changes made in the HBlank Stat interrupt handler could be
reflected in the current line while were intended for the next line.
This is because a line based renderer is used, which renders the
complete line in the hblank period, which in the original Game Boy can
be used to change render settings.