Implement the window internal line counter
The window internal line counter is similar to the global line counter
LY. It starts at 0 and is reset in vblank. It is incremented each time
the window rendering is triggered on a scanline.
The value of the internal line counter is used instead of LY during
window tile fetching.