Add mooneye MBC1 tests as unit tests
Also fix the non-conforming MBC implementation. Previously address bits
which were unused because the address space was not large enough (i.e.
the cartridge did not have enough RAM/ROM to use all address bits) were
not ignored, leading to an out of bounds access. Now those bits are
ignored.
The amount of rom bank numbers was off by a factor of 2.