Add mooneye MBC5 test
The following errors were fixed:
The assertion for the rom address used the ram size instead of the rom
size.
The address bits which are not required to address into a given RAM/ROM
size were not masked.
The MBC5_RUMBLE_RAM cartridge type was not recognized as an MBC5
cartridge.
The ram was enabled for every value with A in the lower nibble, while
the Gameboy CTR says it should only be enabled for value 0x0A.