Precompute tile lines in a cache to improve performance
Two bytes are used to represent an 8x8 tile in the gameboy. For the line
based renderer, we have to get the line of the tile which should be
displayed on the current tile and render it. Instead of calculating the
lines content on the fly, we precompute all possible tile lines made up
of all combinations of two bytes at compile time, indexable by the two
bytes.