Install libgtk in CI
The conan dependency for libgtk does not work yet, since it is only an indirect dependency of this project. Nativefiledialog-Extended depends on GTK. Since NFD-E is not on conan, it is installed using FetchContent. But then it can't find anything installed by conan out of the box. Until that is fixed, install gtk in the system manually