Use nativefiledialog instead of nativefiledialog-extended
Nativefiledialog is available on conan, so the manual dependency management
using FetchPackage is not required anymore. Also the manual install of GTK3.0 is
not required after this change, since conan handles that dependency now.
I initially selected nativefiledialog-extended instead of nativefiledialog since
it has a CMake build and a C++ api. But with conan the build system does not
matter and the C++ api is not much different to the C api (except from manual
memory management, but that can be handled with a smart pointer).