Remove nas dependency
Due to a recent change, the nas recipe is incompatible with clang
(https://github.com/conan-io/conan-center-index/pull/16267#issuecomment-1469824504).
Locally I had the previous version in the cache, but it failed on CI. Since I
dont use nas (it came as a dependency of SDL), configure the dependency away to
be able to build with clang.