Fix pixel index calculation error
In an earlier refactor, the draw_sprites_line function was split into one for
normal and one for tall sprites. The function for tall sprites used the pixel
index calculation routine with a sprite height for normal sprites, meaning an
unsigned underflow would lead to a very large index.
To fix this, pass the sprite height (8/16) to the function.