Add github release action
Currently this only publishes the source and no prebuilt binaries. For that more
dependency management would be required, either all runtime libraries would have
to be installed on the users system or they would have to be bundled or
statically linked.
Also for compatibility with older systems the glibc version of the build node
would have to be taken into account.