Make emulator buildable using the emscripten toolchain
TODO:
- Port file dialog (currently it is just deactivated). Probably needs
another method to select a rom file from the users computer using the
browsers file picker
- Improve ram file implementation: Previously a memory mapped file was
used which implemented saving games (which worked using battery backed
memory on the original hardware). While the current implementation
builds, saving the game is not possible
- While it builds and starts, there are no windows displayed, probably
because the imgui/renderer combo does not work with Emscripten.