Get dev container buildable again
LLVM 19 has uninstallable dependency versions on Ubuntu 22.04. Downgrade to 17 as a solution.
Include-what-you-use cannot be built from source due to a linker error during CMake generation when the C++ compiler is tested. Here, it is unable to link to libstdc++, failing to find a file. This issue might be caused due to a non-standard GCC version being installed. The downside of trying to get the latest tools ;)
Building directly from source without an intermediary stage will bloat the container anyway due to source and build artifacts being included in the final image.