Refactor way to reload games during runtime
The previous solution "worked" by using placement new to reconstruct the Emulator object at the old stack location to reset the state of the emulator when switching games, but could best be described as a hack. It did not call the destructor. It also contained the logical bug of resetting and not restoring the stored callback std::function members in Emulator, causing missing audio.

Reset SDL audio buffers and queues when reloading a game to stop playing old audio as soon as possible. This does not fully prevent a slight audio glitch when switching games, but there is nothing to do to prevent it, see the docs of SDL_ClearQueuedAudio:
"This will not prevent playback of queued audio that's already been sent to the hardware, as we can not undo that, so expect there to be some fraction of a second of audio that might still be heard."