Improve .clang-tidy file readability
The clang-tidy docs say the following about the syntax of a .clang-tidy file:
"Checks: Same as '--checks'. Additionally, the list of globs can be specified as a list instead of a string."
This is wonderful news since it allows me to split up the long unreadable line and make it one line per check. And I even get to comment the list!
CLion does not seem to like the syntax and fails schema validation because it expects a string here. But "clang-tidy --dump-config" reads it without issues.

Deprecated AnalyzeTemporaryDtors was removed (https://github.com/llvm/llvm-project/issues/62020)